@@ -1,37 +1,19 @@
-// Top-level build file where you can add configuration options common to all sub-projects/modules.
-buildscript {
-    ext.kotlin_version = "1.7.0"
-    repositories {
-        maven { url 'https://maven.aliyun.com/repository/google' }
-        maven { url 'https://maven.aliyun.com/repository/public' }
-        maven { url 'https://jitpack.io' }
-        google()
-    }
-    dependencies {
-        classpath 'com.android.tools.build:gradle:7.2.1'
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
-        classpath 'com.github.dcendents:android-maven-gradle-plugin:2.0'
-
-        // NOTE: Do not place your application dependencies here; they belong
-        // in the individual module build.gradle files
-    }
+plugins {
+    id 'com.android.application' version '8.2.2' apply false
+    id 'com.android.library' version '8.2.2' apply false
+    id 'org.jetbrains.kotlin.android' version '1.8.10' apply false
 }
 
-allprojects {
-    repositories {
-        maven { url 'https://maven.aliyun.com/repository/google' }
-        maven { url 'https://maven.aliyun.com/repository/public' }
-        maven { url 'https://jitpack.io' }
-        google()
+task clean(type: Delete) {
+    for (project in allprojects) {
+        delete project.buildDir
     }
 }
 
 ext {
-    compileSdkVersion = 31
-    buildToolsVersion = "28.0.3"
-
-    targetSdkVersion = 31
-    minSdkVersion = 19
+    compileSdkVersion = 33
+    targetSdkVersion = 33
+    minSdkVersion = 21
 
     versionCode = 1
     versionName = "1.0.7"
