@@ -3,21 +3,21 @@ package com.mx.example.app
 import android.content.Intent
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
-import com.mx.example.R
-import kotlinx.android.synthetic.main.activity_main.*
+import com.mx.example.databinding.ActivityMainBinding
 
 class MainActivity : AppCompatActivity() {
+    private val binding by lazy { ActivityMainBinding.inflate(layoutInflater) }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-        expireTestTxv.setOnClickListener {
+        setContentView(binding.root)
+        binding.expireTestTxv.setOnClickListener {
             startActivity(Intent(this, ExpireTestActivity::class.java))
         }
-        threadTestTxv.setOnClickListener {
+        binding.threadTestTxv.setOnClickListener {
             startActivity(Intent(this, MultThreadTestActivity::class.java))
         }
-        delegateTestTxv.setOnClickListener {
+        binding.delegateTestTxv.setOnClickListener {
             startActivity(Intent(this, DelegateTestActivity::class.java))
         }
     }
